🛠 refactor: passport module 제거
